requirements.txt
vpype[all]
vpype-gcode
pybricksdev
